refactor: move hasher to security folder
