+++ src/application/protocols/account/index.ts
@@ -1,3 +1,2 @@
 export * from './add-account-repository'
 export * from './check-account-by-email-repository'
-export * from './hasher'
