+++ src/application/protocols/security/hasher.ts
