+++ src/application/protocols/security/index.ts
@@ -1,0 +1,1 @@
+export * from './hasher'
