+++ tests/application/mocks/account/index.ts
@@ -1,3 +1,2 @@
 export * from './add-account-repository-mock'
 export * from './check-account-by-email-repository-mock'
-export * from './hasher-mock'
